[refactor] categories into categories_encoder and category loss
